code refactoring
- added seperate actions
-  added ping to url method
