--- src/components/home/actionConstants.js
@@ -1,3 +1,0 @@
-export default {
-  UPDATE_NETWORK: "UPDATE_NETWORK"
-};
