+++ src/components/home/reducer/constants.js
@@ -1,0 +1,10 @@
+export const types = {
+  UPDATE_NETWORK: "UPDATE_NETWORK"
+};
+
+export const initialState = {
+  network: null,
+  networkUpdatePending: false
+};
+
+export default types;
