--- src/components/renderer/actionConstants.js
@@ -1,6 +1,0 @@
-// Actions
-export default {
-  UPDATE_DOCUMENT: "UPDATE_DOCUMENT",
-  DOCUMENT_TEMPLATE_REGISTER: "DOCUMENT_TEMPLATE_REGISTER",
-  DOCUMENT_TEMPLATE_SELECT_TAB: "DOCUMENT_TEMPLATE_SELECT_TAB"
-};
