+++ src/reducers/index.js
@@ -1,7 +1,7 @@
 import { combineReducers } from "redux";
 
-import home from "components/home";
-import renderer from "components/renderer";
+import home from "components/home/reducer";
+import renderer from "components/renderer/reducer/document";
 
 export default combineReducers({
   home,
