+++ src/services/web3/getWeb3.js
@@ -4,14 +4,12 @@ import { NETWORK_TYPES, INFURA_PROJECT_ID } from "config";
 let web3Instance;
 
 async function loadWeb3InfuraWebsocket(mainnet = true) {
-  console.log("new demo mainnet", mainnet);
   const net = mainnet ? "homestead" : "ropsten";
 
   return new ethers.providers.InfuraProvider(net, INFURA_PROJECT_ID);
 }
 
 async function loadWeb3Injected() {
-  console.log("new demo injected");
   let { web3 } = window;
   const alreadyInjected = typeof web3 !== "undefined";
 
@@ -23,7 +21,6 @@ async function loadWeb3Injected() {
 }
 
 async function loadWeb3CustomRpc(rpc = "http://localhost:8545") {
-  console.log("new demo rpc");
   return new ethers.providers.JsonRpcProvider(rpc);
 }
 
@@ -69,7 +66,6 @@ async function resolveWeb3(
 }
 
 export function setNewWeb3(t, config) {
-  console.log("new demo");
   return new Promise((resolve, reject) => {
     // Wait for loading completion to avoid race conditions with web3 injection timing.
     // Server-side rendering fails when trying to access window
@@ -94,7 +90,6 @@ export const fetchNetwork = async provider => {
 };
 
 export function getWeb3(t, config) {
-  console.log("trye here");
   if (web3Instance) {
     return new Promise(resolve => {
       resolve(web3Instance);
